editoast: move Ordering into views::operational_studies
