+++ editoast/src/models/scenario.rs
@@ -23,10 +23,10 @@ use crate::error::Result;
 use crate::models::train_schedule::LightTrainSchedule;
 use crate::models::Delete;
 use crate::models::TextArray;
-use crate::modelsv2::projects::Ordering;
 use crate::modelsv2::DbConnection;
 use crate::modelsv2::DbConnectionPool;
 use crate::tables::scenario;
+use crate::views::operational_studies::Ordering;
 use crate::views::pagination::Paginate;
 use crate::views::pagination::PaginatedResponse;
 
@@ -231,7 +231,7 @@ pub mod test {
     use crate::models::List;
     use crate::models::Retrieve;
     use crate::models::Timetable;
-    use crate::modelsv2::Ordering;
+    use crate::views::operational_studies::Ordering;
 
     #[rstest]
     async fn create_delete_scenario(db_pool: Arc<DbConnectionPool>) {
