+++ editoast/src/modelsv2/study.rs
@@ -15,7 +15,6 @@ use utoipa::ToSchema;
 
 use crate::error::Result;
 use crate::models::List;
-use crate::modelsv2::projects::Ordering;
 use crate::modelsv2::projects::Tags;
 use crate::modelsv2::Changeset;
 use crate::modelsv2::DbConnection;
@@ -23,6 +22,7 @@ use crate::modelsv2::DbConnectionPool;
 use crate::modelsv2::Model;
 use crate::modelsv2::Row;
 use crate::modelsv2::Save;
+use crate::views::operational_studies::Ordering;
 use crate::views::pagination::Paginate;
 use crate::views::pagination::PaginatedResponse;
 
@@ -127,8 +127,8 @@ pub mod test {
     use crate::models::List;
     use crate::modelsv2::DeleteStatic;
     use crate::modelsv2::Model;
-    use crate::modelsv2::Ordering;
     use crate::modelsv2::Retrieve;
+    use crate::views::operational_studies::Ordering;
 
     #[rstest]
     async fn create_delete_study(
